requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mako==1.1.6
markupsafe==2.1.0
passlib==1.7.4
psycopg2==2.9.3
pydantic[dotenv]==1.9.0
pyjwt==1.7.1
python-dotenv==0.19.2
sniffio==1.2.0
sqlalchemy==1.4.32
starlette==0.16.0
typing-extensions==4.1.1
uvicorn==0.15.0